requirements.txt
playwright
pytest-playwright
pytest-asyncio
